heart_disease_dashboard/requirements.txt
pandas
panel
numpy
matplotlib
sqlalchemy
python-dotenv
pymysql
plotly
seaborn
